+++ lib/src/modules/elss/common_models/elss_device_details_response.dart
@@ -75,6 +75,9 @@ class DeviceDetailsData {
   @JsonKey(name: "st")
   List<String>? stockTags;
 
+  @JsonKey(name: "rs")
+  int? rubbingOrGlassChange;
+
   DeviceDetailsData(
     this.deviceBarcode,
     this.requestReason,
@@ -95,6 +98,7 @@ class DeviceDetailsData {
     this.isMarkRepairedDevice,
     this.serialNumber,
     this.stockTags,
+    this.rubbingOrGlassChange,
   );
 
   static DeviceDetailsData fromJson(Map<String, dynamic> data) => _$DeviceDetailsDataFromJson(data);